requirements.txt
huggingface-hub==0.23.0
gradio==4.19.2
imageio[ffmpeg]>=2.9.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
